requirements.txt
-i https://pypi.org/simple
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
chardet==3.0.4
idna==2.8
multidict==4.5.2
pyyaml==5.1.2
telepot==12.7
urllib3==1.25.6
yarl==1.3.0
zabbix-api==0.5.4
